requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sentence-transformers
faiss-cpu
numpy
google-adk
python-dotenv
